fix(client): resolve Firebase env var errors by moving server types to shared files
